XIVY-9397 Always provide root-relative url as originalUrl
So that it is always possible to redirect from anywhere in
the hole ivy system to the original url. Because I want
to pick-up this original url for OAuth2 logins. So that
I can redirect to this URL after login, which will be
redirected from a complete other place (OAuth2Servlet).