XIVY-11246 Fix npe after user switch if orignal url is case/task detail
The problem is not that the case/task id is not on the parameter, the
problem is that if you do a redirect to the login page, then a POST
request is first triggered (without any query params). The original URL
is then evaluated from this POST request instead of the inital page
request.