XIVY-16798 Fix NPE on task redirect
If a POST ajax request is happening on the task detail page, and after
that the task is directly started, the originUrl will be evaluated based
on this POST request. The problem here is that this request does not
include the query parameters, so the task id is no longer there and
therefore there is no selected task.